+++ querydsl-sql-r2dbc-spring/src/test/java/com/github/xuse/querydsl/sql/r2dbc/service/SpringService.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.github.xuse.querydsl.lambda.LambdaHelpers;
 import com.github.xuse.querydsl.lambda.LambdaTable;
-import com.github.xuse.querydsl.r2dbc.core.R2dbcFactory;
-import com.github.xuse.querydsl.r2dbc.core.R2dbcFactory.R2Fetchable;
+import com.github.xuse.querydsl.r2dbc.R2dbcFactory;
+import com.github.xuse.querydsl.r2dbc.R2dbcFactory.R2Fetchable;
 import com.github.xuse.querydsl.sql.r2dbc.entity.Foo;
 import com.github.xuse.querydsl.sql.routing.RoutingStrategy;
 import com.github.xuse.querydsl.util.StringUtils;
